+++ lib/get-schema-map/get-schema-map.d.ts
@@ -3,4 +3,4 @@ import { SchemaNode } from 'dgraph-js';
 export interface ISchemaMap {
     [predicate: string]: SchemaNode;
 }
-export declare function xGetSchemaMapNow(dgraphClient: dgraph.DgraphClient): Promise<ISchemaMap>;
+export declare function xGetSchemaMapTxn(dgraphClient: dgraph.DgraphClient): Promise<ISchemaMap>;