+++ lib/set-schema-now/set-schema-now.d.ts
@@ -1,2 +1,0 @@
-import * as dgraph from 'dgraph-js';
-export declare function xSetSchemaNow(schema: string, dgraphClient: dgraph.DgraphClient, _dgraph?: typeof dgraph): Promise<dgraph.Payload>;