+++ lib/upsert-map/upsert-map.d.ts
@@ -5,4 +5,4 @@ export interface IObjectMap {
 export interface IUidMap {
     [key: string]: string;
 }
-export declare function xUpsertMapNow(queryFn: any, data: IObjectMap, dgraphClient: dgraph.DgraphClient, _dgraph?: typeof dgraph): Promise<IUidMap>;
+export declare function xUpsertMapTxn(queryFn: any, data: IObjectMap, dgraphClient: dgraph.DgraphClient, _dgraph?: typeof dgraph): Promise<IUidMap>;