+++ lib/upsert/upsert-fns/basic-equality-upsert-fn.d.ts
@@ -1,2 +1,2 @@
-import { IUpsertFnReturnValues } from '../upsert-now';
+import { IUpsertFnReturnValues } from '../upsert';
 export declare const basicEqualityUpsertFn: (searchPredicates: any) => (node: any) => IUpsertFnReturnValues;